glq documentation examples
